+++ internal/controller/repo.json
@@ -1,0 +1,1 @@
+{"b8da4f2d":"www.ya.ru","ba980180":"https://practicum.yandex.ru"}
